Update index.js
We are only downloading the base image if the `options` param explicitly includes the option `{prepareBaseImage: false}`. Otherwise, it will omit downloading the base file, even if the helper configuration includes the prepareBaseImage option.

This change allows us to default to the value in the config file if the method doesn't receive that param.